Add version of the RxPlayer in the debug element
I like relying on our debug element feature when debugging elements and
other users seem to also find that element useful, sharing to us screen
captures/recording of it.

In that context, it could be good to also display the current RxPlayer
version in it, so we can quickly know which version is currently relied
on.