MULTI_THREAD: Fix onmessageerror being undefined
We recently defined a `onmessageerror` global callback on the
WebWorker-side of the `MULTI_THREAD` feature, thinking that it was
defined, like the `onmessage` callback, as long as WebWorker were
accessible.

Turns out that was not the case at all, and the multithreading mode was
breaking on older device. Oops, sorry about that.

As it is not required, we can just access it through our `globalScope`
shortcut to the global object. This way, it will just define a callback
which will never be called if it does not exist.