Test: Run integration tests also on Edge and Safari
Until now, CI/npm scripts only runs our integration tests on Chrome and
Firefox.

This is a proposal to also run them on Edge and Safari.

There have been very few browser-dependent issues found with integration
tests for now but it may still be useful and wasn't too hard to do.
Testing on both edge and Safari also allows to run a lot of MS-specific
and Webkit-specific workarounds which may allow to better detect
browser-specific regressions.

I tested it and it seemed to work.

Issues with this is that we have more chance to be annoyed by flaky tests
(which would be better fixed by updating the tests anyway) and that CI
steps would take longer. We may want to split CI scripts per-browser for
that last part? To see.