Do not communicate needlessly the media element to multiple modules
While rebasing our Proof-of-concept for content preloading, I noticed
that multiple modules asked for the `HTMLMediaElement` to be provided to
them as argument despite the fact that most of them also rely on a
`PlaybackObserver` which fulfill most of its roles.

I decided to just remove that unneeded dependency for them.

A longer-term end goal might be to make `PlaybackObserver` the
preferred interface to the `HTMLMediaElement` in general (and
perhaps rename it), a bit like the `SourceBufferInterface` is
for `SourceBuffer` objects. We're not too far from this.