Make the `MULTI_THREAD` feature not experimental anymore
I propose here to make our `MULTI_THREAD` feature, one the main feature
brought in our `v4.0.0`, a stable (i.e. not `experimental` anymore) one.

At Canal+ Group, it has been the main way to rely on the RxPlayer for
more than a year (with gains in terms of adaptive stability and
performances), and with #1627, we're in the process of making the
multi-thread code path the main one that will be relied on (with an
adaptation to support the default monothread one).

Removing the `experimental` label would encourage applications to rely
on this feature more, which would bring both its advantages to them and
align them more with what we're used to internally at Canal+.

We've never changed its API but still kept its "experimental" label
mainly to let us update its API, especially for the following issues:

1. For now it only supports DASH streaming, which is directily included
   with the worker code.

   The issue with supporting more streaming protocols is that we don't
   want to include those other rarely-used ones in every applications
   using the feature, and that we're still unsure how a feature
   switching concept could be relied on together with `MULTI_THREAD`
   (as the feature code to inject should be sometimes injected to the
   worker which runs in another thread - adding communication issues).

2. For a long time, I was not sure if we should directly include the
   "embedded worker" file in the `MULTI_THREAD` feature (as 95% of
   applications should want that) or if we let this as a supplementary
   step to let the application do more advanced things if it wants
   (worker transpiled to ES5, stored on a CDN somewhere and loaded
   lazily etc.).

   For now I chose the more flexible one as I think that this
   supplementary step is not a huge added complexity to its API.

3. It has some limitations (DASH-only, no `manifestLoader` or
   `segmentLoader`, [platform compatible to
   WebWorkers](https://caniuse.com/?search=webworker)), which is why we
   often advised people to rely on both `MULTI_THREAD` and on the
   main-thread `DASH` feature as a security fallback if those
   limitations aren't respected.

   This means more complexity in how the player may run and much more
   code being loaded (for what is most times unnecessarily
   duplicated code).

Keeping the API experimental would let us maybe find a way to fix or reduce
those issues at some point, but I think the gain of making it stable
(encouraging more people to use it and to report issues/needs on this)
make it up for the issues of doing it (less freedom to update its API).