CancellationError: Add descriptive task name
I've had an idea while reading some debugging stuff online that may or may not be a good idea, but I'm proposing it here anyway.

Context: `CancellationError` have a generic error `message`
-----------------------------------------------------------

The `TaskCanceller` is our homemade abstraction to handle task cancellation (its API is similar to the `AbortController` web API).
With it, when a cancellable task is cancelled, it usually will throw or reject a special `CancellationError` `Error` with the message `"This task was cancelled"`.

The thrown message is very generic and could be thrown by any place in the code, making debugging when starting from a `CancellationError` hard.

Note that those errors should never be seen by the application: we try to internally handle all of those. Though we've seen some very rare cases where they have "leaked".
We've had (maybe once or twice, so not that frequent) reported cases where such errors has been seen by other developers in some impossible-to-reproduce scenario, in which case it's hard to understand what happened.

More frequent are cases where we are debugging strange behavior linked to task cancellation (or the absence of it), here also we could do much better by improving on a TaskCanceller's tracing.

Solution: indicate unique name for each task and log it
-------------------------------------------------------

So basically what I'm proposing here is to encourage RxPlayer maintainers to declare a unique "name" for a task when creating either a `TaskCanceller` or a `CancellationError`.

That name will be indicated in the corresponding `CancellationError`'s message.

More interestingly, I now log every case where a `CancellationError` is instantiated (and not thrown! As it is easier to write :/). This leads to more logs, but it's IMO more informative on what goes on inside the RxPlayer.

The downside is that a lot of cancellations are done when stopping or reloading a content, leading to a good majority of uninteresting logs (unless we're trying to find a leak!) when those actions are performed. This also adds more arguments for the few outsiders which tell us that we output too much logs in debug mode!