[MULTI_THREAD] Communicate to `/init` whether we're relying on MSE-in-worker
Built on #1664 (the PR, not the other thing).

While exchanging in that PR with @Florent-Bouisset, we noticed that the
choice of whether we rely on mse-in-worker was done by the Public API
part of the code, which then directly told our WebWorker to use that
feature without telling any other part of the code.

However, that PR brought the need to have that information inside the
`ContentInitializer` - so that it knows whether a given codec is
supported (for cases, only seen on Edge for now, where codec support is
different depending on if MSE is relied on inside a worker or on main
thread).

Even without this, I thought that our `ContentInitializer`, as a global
controlling module, should probably architecturally be in the know of
whether MSE API are called in-worker or in-main-thread.

So I propose this modification, where the information on whether
mse-in-worker should be used is both communicated to the
`ContentInitializer`, but is also set per-content (we don't need that,
but it's actually both simpler and more powerful that way so why not).

This also opens the way for an API to control this, but I did not add
one for now.