DRM: Fix `KEY_UPDATE_ERROR` by not translating it into a `KEY_LOAD_ERROR`
We had kind of a dumb mistake where an error arising at
license-communication time was falsely categorized as an error at
license-fetching time.

This is because we inadvertently also catched errors from the
`MediaKeySession.prototype.update` EME call where we only intended to
catch errors at the higher-up `getLicense` (the RxPlayer API) level.

The fix to that issue is relatively straigtforward.