Avoid error log when stopping a stream with a pending GC buffer removal
Fixes #1668

Basically, if an application stopped playback while buffer-removal due
to `maxBufferAhead` / `maxBufferBehind` options were set, an error log
would be outputed:
```
Could not run BufferGarbageCollector: This task was cancelled.
```

This log in this case is unnecessary (the behavior is normal and wanted)
and its verbosity could make people think there was an issue.