Add information on key system to an `EncryptedMediaError`
We worked with an application that wished, for resilience reasons, to
be able to switch from a key system to another (e.g. widevine to
playready) when specific encryption errors occured and when both key
systems were available.

Previously they used our `getKeySystemConfiguration` API to know about
the current key system used.

But when processing an error, it might be too late (e.g. the key system
linked to the RxPlayer might already has been reset due to the error) or
too soon (the issue happened before we could confirm the current key
system through that API).

Because it seems logical to me as an API, I propose here to add the fields
`keySystem` (the name/RDN of the key system as advertised by
`MediaKeySystemAccess.prototype.keySystem`) and
`mediaKeySystemAccessConfiguration` (the configuration as advertised by
the `MediaKeySystemAccessConfiguration.prototype.getConfiguration()`) on
all `EncryptedMediaError` instances when it is known (so on all errors
but `INVALID_KEY_SYSTEM`, `INCOMPATIBLE_KEY_SYSTEMS` and
`MEDIA_IS_ENCRYPTED` errors as they all happen before the
`MediaKeySystemAccess` has been obtained).

Consequently, an application can then write code like:

```js
player.addEventListener("error", (error) => {
  if (error.type === "ENCRYPTED_MEDIA_ERROR") {
    const { keySystem } = error;
    if (keySystem !== undefined && keySystem.indexOf("playready") >= 0) {
      // we were having the issue while using playready
    }
  }
});
```