Begin to formalize logs
This is a Work-In-Progress and a proposal to add a little "formalization"
for our logs, as those became our central focus for debugging complex
issues.

There's already an implicit log API we have to maintain as we rely on a
given format in our [RxPaired](https://github.com/canalplus/RxPaired)
live debugger but this was just enforced by not changing too much the
lready-present logs.

We also noticed that multiple partners outside of canal+ where using
RxPaired and looking at our logs for debugging on their side first.

---

Consequently I tried a strategy which attempts to:

- Facilitate the readability of our logs by people relying on the
  RxPlayer. This is here done in two ways:

  1. The "namespace" (first part of the log, before `:`) has to be taken
     from a smallish list of authorized namespaces, enforced by
     TypeScript. E.g. `ABR` for adaptive, `DRM` for decryption, `mse`
     for MSE-linked code, `SF` for segment fetching code etc.

     The end idea would even to allow clear documentation for those
     namespaces so that people using the RxPlayer have a vague idea of
     what the various messages we log are about.

     It also helps by enforcing more stability (NOTE: I do not consider
     the namespace as part of our API, but we should always have a very
     good reason to break it) for external tools exploiting those.

  2. When there's data associated with the log, it was previously just
     appended to the message without description (e.g. some numbers
     after a vague message about ABR).

     Now the preferred syntax is under a JS Object which is then written
     as `key1=value1 key2=value2` In our log messages. This helps
     people (including us here :p) quicly understand what those values
     are about.

- Try to put some more stability in our log for our external tools like
  RxPaired, through the same aforementioned strategies

---

The main downside I see is that adding logs needs a little more looking
around to get the namespace and format right.

Thoughts?