Use comments to delimit environment-specific code
The need: environment-specific code
-----------------------------------

We sometimes need to define some code that only run in development builds.

The idea is that we want a RxPlayer that easily breaks when anything unexpected happen
only when we locally test the RxPlayer (so we can detect it directly), yet with production
builds running on a customer's device we want to stay resilient even in completely
unexpected scenarios.

Usually for JS developers, this has been done through compile-time constants, e.g. a
`__DEV__` boolean defined globally and replaced at build-time by the build tools
themselves through configuration. Most JS applications rely on a "bundler" tool like
`webpack`, `rollup`, `vite` or `esbuild` to produce a JS bundle ready to be deployed in an
actual website, and those bundler tools all support some compile-time constant system.

But the RxPlayer is a library that mostly rely on TypeScript without any bundler. And
sadly TypeScript doesn't have a compile-time constant system and doesn't appear ready to
include one.

Previous solution: TypeScript's `const enum`
--------------------------------------------

Because we rely on TypeScript without any bundler for our regular builds, we until now
resorted instead to a trick where we added `const enum` in our code.

`const enum` are like TypeScript `enum` (it lists a set of constants) but with a nice
supplementary feature: where an `enum` compiles down to supplementary JS Objects with
property lookups when they are accessed, a `const enum` is totally erased, and the linked
lookups are replaced by the corresponding constants in the built files.

We relied on `const enum` both to easily update at a single place the wanted environment
(two possible files `dev.d.ts` and `prod.d.ts` that would define the same `const enum` but
with only one of them selected at build-time) so the corresponding code could run or not
(if it was just for this `enum` could also do the trick), but also so unreachable code can
be sufficiently simple to determine to be optimized away by a minifier.

For example, when needing to define code only intended for development builds, we would
write (this ugly code):

```js
// (Where `__ENVIRONMENT__` is defined as a const enum somewhere else)
if ((__ENVIRONMENT__.CURRENT_ENV as number) === (__ENVIRONMENT__.DEV as number)) {
  // Some dev-only code
}
```

Gives once-compiled in production mode something like:

```js
if (0 /* __ENVIRONMENT__.CURRENT_ENV */ === 1 /* __ENVIRONMENT__.DEV */) {
  // Some dev-only code
}
```

And most minifiers are today able to remove that impossible-to-run code.

My opinion with the previous solution
-------------------------------------

This solution is not easy to understand, rely on implicit side-effects of multiple tools
(TypeScript and the behavior associated to `const enum`, the fact that minifiers are smart
enough, the ability to select different `const enum`-defining files at build time) that
could break in the future.

This always has been a subpar solution, but I'm thinking it only becomes worse with time:

- TypeScript becomes smarter and smarter, which led us to the necessity of adding first
  the equality, then the `as number` specifiers.

- `const enum` were not added as a compile-time constant structure, are not documented as
  such,
  [and are even not recommended by TypeScript themselves](https://www.typescriptlang.org/docs/handbook/enums.html#const-enum-pitfalls).

- [Node.js began automatically stripping typescript types](https://nodejs.org/api/typescript.html#type-stripping),
  and we saw other similar efforts of minimal type-stripping (without type-checking) which
  do not support the more advanced and out-of-favor TypeScript features like `enum` (and I
  suppose also `const enum`) constructs which needs more intelligence than just stripping
  supplementary syntax.

- As we added more lint and type-checking in the project, we also had to add multiple
  `tsconfig.json` files, where half of them (the `.dev.json` ones) are only there because
  of that trick. This makes the project harder to understand for maintainers.

Current solution: "Code block comments"
---------------------------------------

I propose here a different solution where such specific code blocks are delimited by
specific comments.

For example:

```js
// ## START CODE BLOCK: DEBUG-BUILD-ONLY
if (someImpossibleCondition) {
  throw new Error("An impossible situation arised :O!");
}
// ## END CODE BLOCK: DEBUG-BUILD-ONLY
```

At build time, a homemade script will read all input files and automatically strip the
code we do not need before feeding the file to TypeScript / the bundler.

I measured the time lost by that stripping logic and it was very negligible: ~1.5ms on my
PC for what account to ~7 seconds of build time.

I'm however not totally convinced.

Pros:

- We remove the reliance on the ugly / poorly-adapted `const enum` trick

- We can remove the large number of tsconfig.json config files we have and just set one
  intended for type-checking, linting, LSP and such general-purpose development tools (and
  not for builds).

- We do not rely on the application's minifier to remove unreachable code, it is already
  removed at build-time.

- Relatively simple: We don't have to plug ourselves on TypeScript's AST, we don't have to
  understand JS semantics, the same stripping logic can be repeated both for TypeScript and
  type-stripping bundlers (esbuild here).

- Should stay low-maintainance: even for cases where we want to change build dependencies,
  pre-processing the input files should always be relatively easy to do.

- Alternatives all have flaws

Cons:

- Our build logic has a supplementary "code-stripping" step that could be easy to miss
  when changing the build logic / when adding a new build system.

- It relies on TypeScript's node.js API which is poorly documented. Moreover it weirdly
  seemed to lead to ~7% worse build performance (6300 ms for the cli vs 6750 ms for the
  node.js API) but perhaps it's due to a poor comprehension of the API on my part.

  I also had to glue things together, like re-feeding TypeScript its own produced
  diagnostics so that it can format them for stdout, without really understanding why
  TypeScript would need me to do that (I suspect a simpler API to be available).

- TypeScript will not always see what we see in our editor when compiling: Possibility to
  have build-time only TypeScript issues.

- Developers could be surprised that some code disappears once built, whereas an in-JS
  identifier is more visible.

- Mistakes are not checked automatically, we have to be careful of typos when reviewing
  (we could limit this with more advanced stripping script perhaps?).

- Sometimes TypeScript or linters like to be smart, and don't understand how situations
  are possible when it is only when stripping code. That issue also existed with the
  previous `const enum` solution but now there are some more cases, most
  notably with rules about not using defined variables and about unreachable
  code.