prober: throw if the HDCP API is not available
ProbeHDCPPolicy can either, succeed with "Supported" | "NotSupported" | "Unknown" or reject.
Success with "unknown" or reject both occurs when the navigator API failed to execute and the support
of HDCP cannot be determined. To standardize the behavior of this function, the "Unknown"
return is removed and replaced by a promise rejection.