Add package_live_content.sh script
This commit improve on our test capabilities, by generating contents
at test-time, optionally live and encrypted with clearkey encryption.

It relies on [`ffmpeg`](https://www.ffmpeg.org/) and on the
[`shaka-packager`](https://github.com/shaka-project/shaka-packager) to do so.
The former is ubiquitous and easy to install, the second
[provide binaries](https://github.com/shaka-project/shaka-packager/releases) for linux and
OSX (yes, there's [particularly one](https://www.haiku-os.org/) sadly missing here).
[We're already optionally loading binaries](https://github.com/canalplus/rx-player/blob/dev/scripts/install_rust_toolchain.sh)
at build-time (mainly for one-time contributors and the CI) with success so I guess this
is OK to add those as dependencies for integration testing-time?

I used predominently noone's favorite language, ̶J̶a̶v̶a̶S̶ Bash, because of the ease to just
call local binaries in it.

---

Doing this allows to resolve a blind spot we have in our tests: live contents. It can also
reinforce our test capabilities on our most frequent source of issues: DRMs.

If it turns out to be working as intended, it can also be relied on as an alternative from
https://github.com/canalplus/rx-player/pull/1478 (which I still think should be merged),
which patched browser APIs (MSE+EME+HTML) to let us feed whatever fake content and test
only the RxPlayer's behavior.

---

Also while doing this I realized that it also can be used for local tests,
where we do not have to search the web for test contents with specific
characteristics, we can just try to package it ourselves first.