PRETTIER RANDOMLY DECIDED TO REMOVE SOME RANDOM-ASS BLANK LINES AND DO A HUGE DIFF
I thought adopting prettier would just lead to a stable and approachable
style.

It turns out that doing that means a lot more style bikeshedding, just
outsourced to unknown people on the internet.

Now that we adopted it, changes that I would guess nobody sane actually
cares about - like the presence of an initial blank line or not in a
markdown list - randomly change at some point.

It's just annoying to have conflicts because a random person at prettier
decided that the old one was ugly or something (breaking news: nobody
cared about list line breaks in md, it just bothers people to no end to
have all those modifs for no gain)