Fix persistent session loading when content has no key id
We saw issues with some contents from a partner using persistent
licenses.

After investigations, I noticed that the stored metadata linked to
persisted licences could be corrupted if, and only if, key ids are not
known in advance through the Manifest nor through segments.

Thankfully this is a very rare occurence, itself happening also only in
the rare scenario where persisted sessions are used - but we did actually
saw it with content from a partner (italian one).

---

The root cause was due to a simple JS gotcha: the result of an Array's
`every` method (which returns `true` if all elements from an array
follow the given predicate) when the array it is applied to is empty.

Turns out it returns `true` (it even kind of makes sense when you think
hard about it :p!).

Yet we were using it in the following type of condition:
```js
if (currentKeyIds !== undefined) {
  const isCurrentPersistentLicenseCompatible = currentKeyIds.every(keyId => {
    return storedEntry.keyIds.includes(keyId);
  });
}
```

So here, if `currentKeyIds` is defined but empty, it would return `true`
for any `storedEntry`, even ones linked to a license that do not apply
to the current content.

Note that we have a valid fallback in cases where keyIds are unknown,
which is to check whole PSSH instead.
Here we should go through this strategy instead - but we didn't consider
the "empty keyIds array" as equivalent to "no known key id". Now we do.

Also a more efficient solution could also be to tell those partners to just
advertise their key id like everybody else - which also leads to more
performance-linked tricks elsewhere. Though we already tried with
no success.