Fix running of the Edge browser on the CI.
The integration tests fail in the CI since a few days (first seen
wednesday) when it's turn to run the Edge browser.

As the error looks like that:
```
2025-07-18T09:33:55.879Z ERROR edgedriver: Failed to download Edgedriver: request to https://msedgedriver.azureedge.net/138.0.3351.77/edgedriver_linux64.zip failed, reason: getaddrinfo ENOTFOUND msedgedriver.azureedge.net, trying alternative download URL...
2025-07-18T09:33:56.681Z ERROR edgedriver: Failed to download Edgedriver: request to https://msedgedriver.azureedge.net/138.0.3351.77/edgedriver_linux64.zip failed, reason: getaddrinfo ENOTFOUND msedgedriver.azureedge.net, trying alternative download URL...
```

It seems to be due to some stale URL when retrieving "edgedriver", so I guess
something to do with webdriverio (the dependency we use to be able to
run our tests on actual browser implems) and its dependencies.
_Note: we don't even need what edgedriver [seems to be about](https://learn.microsoft.com/en-us/microsoft-edge/webdriver/?tabs=c-sharp),
we just want to run some static page with JS in various browsers. For
now, none of our tests interact with the browser beside running some JS
on it._

It seems to be linked to the https://github.com/webdriverio-community/node-edgedriver/issues/500
issue, which has been fixed here https://github.com/webdriverio-community/node-edgedriver/pull/501
very recently.

That `edgedriver` is a transitive dependency of `webdriverio`:
```
└─┬ webdriverio@9.18.1
  └─┬ @wdio/utils@9.18.0
    └── edgedriver@6.1.2
```

Here I just updated webdriverio as a whole anyway.