feat: add disableAudioTrack API
We previously never added the `disableAudioTrack` API for two reasons:

1. It's often much less needed than a `disableTextTrack` API and than a
   `disableVideoTrack` API (second one for "audio-only")

2. It was unclear what behavior the RxPlayer should have if both audio
   and video were disabled

Since @Florent-Bouisset work on the new `onAudioTracksNotPlayable` /
`onVideoTracksNotPlayable` options however, we do have a concrete way to
deal with the absence of both audio and video: the `noPlayableTrack`
`error` event.

So no that there's no real barrier to add that API, I'm adding it.

---

This does not seem to really be needed right now by the different
applications we work with, I just wanted to add that API now because it
facilitates the writing of edge-cases-tests: I especially wanted to add
a test for a pretty rare situation where there's no audio selected yet
we're fallbacking from all video tracks due to DRM constraints.

I'm doing this now because now that advanced DRM integration tests are
merged, we can begin to add tests for all those advanced edge cases
scenarios to ensure we're keeping them under control.