Thumbnails: Store pending thumbnail request
When I demoed how DASH thumbnail worked in the RxPlayer demo, I was
unhappy to realize that there was a noticeable delay when quickly moving
the mouse over the seekbar.

This is both because we have a timer in our demo of a few milliseconds
before doing the actual thumbnail request, and even much more importantly
because the RxPlayer cancel the last thumbnail request when the
application decides to fetch one from a new position instead.

Yet positions that are sufficiently close (which is frequent when moving
the mouse over a seekbar) are often going to lead to the same thumbnail
request. Here cancelling the last one is unfortunate: we should reuse
the same request.

---

We already have a thumbnail cache but sadly enough it is only populated
once the request is done. If the request is still pending and an
application ask a thumbnail for a new position, we're going to run a new
request even if it leads to the same thumbnail.

Our "last loaded" cache was in main thread, which does not even know the
relation between the wanted time and the corresponding thumbnail, so this
new logic cannot be added there.

Instead I added another level of "cache" in our `ThumbnailFetcher` (in
`core/fetchers`) for the pending request only. The result is much more
satisfying.