Smooth: migrate from DOMParser API to our txml fork
This commit removes the need to have DOM API present when parsing a
Smooth Manifest, by relying on our full-JS tXml lib instead for XML
parsing.

This is already what we do for DASH contents and it mainly will allow to
play Smooth content in a WebWorker context, such as in our
multithreading mode.

Now that something like #1719 is under way, Smooth is the last transport
which needed effort to be ported to run in multithread mode.

Porting it means there will be no feature difference between main and
multithread modes.
This is a nice milestone but will also simplify the documentation of
that complex feature, by removing a key limitation from it.

This commit does not add the Smooth feature to multithread mode yet, but
doing that should be really easy to do (just one line in `src` once #1719
and #1723 are merged).