Update Uint8Array usage so TypeScript's happy
I was updating some dependencies during some project maintainance, but
after updating TypeScript to 5.9.2 (from 5.8), I noticed that a high
number of typing issues appeared.

This appears to be from the fact that now Uint8Array may be generic over
the type of ArrayBuffer it is a representation of (`ArrayBuffer` or
[`SharedArrayBuffer`](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/SharedArrayBuffer))
and that an Uint8Array backed by a `SharedArrayBuffer` is not
compatible to a `BufferSource` type, the type we rely on the most when
calling EME/MSE API or dealing with binary data in general.

In all cases currently, we probably always want an `ArrayBuffer`, as we
don't have any use for a `SharedArrayBuffer` for now. Yet I:

-  didn't want to close the door on using `SharedArrayBuffer`. We're now
   multi-threaded and could at one point decide that using one for some
   binary data is the better thing to do.

   So most utils that could work for both should still work for both.
   I often had to add generics to those to link their inputs to their
   outputs when it made sense to not "lose" an inputed `ArrayBuffer`
   when calling them.

-  Sadly, one API could optionally ask for `Uint8Array` as input:
   the `segmentLoader` `loadVideo` options.

   To not break the API, I decided to detect if what was inputed was
   backed by an ArrayBuffer or not. In the very unlikely event where it
   was not, I copy it (`slice` API) so it's backed by one.